// hub/frontend/lib/api/generated/model/actorSummary.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.75
 */
import type { ActorSummaryType } from './actorSummaryType';

export interface ActorSummary {
  avatarUrl?: string;
  htmlUrl?: string;
  id: string;
  login: string;
  name?: string;
  type: ActorSummaryType;
  url?: string;
}